create scores controller for managing scores CRUD
