+++ spec/controllers/scores_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe ScoresController, type: :controller do
+
+end
